BWA and BLAT integrated successfully
final copy.  should be bug-free.

+++ MetaPro.py
@@ -210,6 +210,7 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
     gene_annotation_BWA_label           = "gene_annotation_BWA"
     gene_annotation_BWA_pp_label        = "gene_annotation_BWA_pp"
     gene_annotation_BLAT_label          = "gene_annotation_BLAT"
+    gene_annotation_BLAT_cleanup_label  = "gene_annotation_BLAT_cleanup"
     gene_annotation_BLAT_cat_label      = "gene_annotation_BLAT_cat"
     gene_annotation_BLAT_pp_label       = "gene_annotation_BLAT_pp"
     gene_annotation_DIAMOND_label       = "gene_annotation_DIAMOND"
@@ -727,6 +728,8 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
         if read_mode == "paired":
             sections.extend(["pair_1", "pair_2"])
         sample_job_flag = True
+        missed_jobs_list = []
+        #job_name = "BLAT_single_job"
         for section in sections:
             for split_sample in os.listdir(os.path.join(gene_annotation_BWA_path, "final_results")):
                 if(split_sample.endswith(".fasta")):
@@ -734,9 +737,10 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
                     file_tag = os.path.splitext(file_tag)[0]
                     full_sample_path = os.path.join(os.path.join(gene_annotation_BWA_path, "final_results", split_sample))
                     for fasta_db in os.listdir(paths.DNA_DB_Split):
-                        if fasta_db.endswith(".fasta") or fasta_db.endswith(".ffn") or fasta_db.endswith(".fsa") or fasta_db.endswith(".fas") or fasta_db.endswith(".fna") or fasta_db.endswith(".ffn"):
+                        if fasta_db.endswith(".fasta") or fasta_db.endswith(".ffn") or fasta_db.endswith(".fsa") or fasta_db.endswith(".fas") or fasta_db.endswith(".fna"):
                             job_name = "BLAT_" + file_tag + "_" + fasta_db
                             BlatPool.apply_async(commands.create_and_launch,
+                            #BlatPool.apply_async(commands.launch_only,
                                 args=(
                                     gene_annotation_BLAT_label,
                                     commands.create_BLAT_annotate_command_v2(gene_annotation_BLAT_label, full_sample_path, fasta_db),
@@ -744,23 +748,46 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
                                     job_name
                                 )
                             )
+                            time.sleep(0.0003)
                             if(sample_job_flag):
                                 print("saving 1 job for sampling:", job_name + ".sh")
                                 sample_job_flag = False
                             else:
                                 print("for file explosion reasons, removing:", job_name +".sh") 
-                                os.remove(os.path.join(gene_annotation_BLAT_path, job_name + ".sh"))
-                            
+                                if(os.path.exists(job_name + ".sh")):
+                                    os.remove(os.path.join(gene_annotation_BLAT_path, job_name + ".sh"))
+                            #    else:
+                            #        missed_jobs_list.append(job_name)
+        print(dt.today(), "final BLAT job removal")
+        for item in os.listdir(gene_annotation_BLAT_path):
+            if(item.endswith(".ffn.sh")):
+                if(os.path.exists(item)):
+                    os.remove(item)
+        #for item in missed_jobs_list:
+        #    if(os.path.exists(item)):
+        #        os.remove(item)
+            
         BlatPool.close()
         BlatPool.join()
 
-        sample_job_flag = True
+        
+        write_to_bypass_log(output_folder_path, gene_annotation_BLAT_label)
+        
+        
+    if check_bypass_log(output_folder_path, gene_annotation_BLAT_cleanup_label):
+        for item in os.listdir(gene_annotation_BLAT_path):
+            if(item.endswith(".ffn.sh")):
+                job_path = os.path.join(gene_annotation_BLAT_path, item)
+                os.remove(job_path)
+        write_to_bypass_log(output_folder_path, gene_annotation_BLAT_cleanup_label)
+        
+    if check_bypass_log(output_folder_path, gene_annotation_BLAT_cat_label):
         for split_sample in os.listdir(os.path.join(gene_annotation_BWA_path, "final_results")):
             if(split_sample.endswith(".fasta")):
                 file_tag = os.path.basename(split_sample)
                 file_tag = os.path.splitext(file_tag)[0]
                 full_sample_path = os.path.join(os.path.join(gene_annotation_BWA_path, "final_results", split_sample))
-                job_name = split_sample + "_cat"
+                job_name = file_tag + "_cat"
                 process = mp.Process(
                     target=commands.create_and_launch,
                     args=(
@@ -770,22 +797,14 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
                         job_name
                     )
                 )
-            if(sample_job_flag):
-                print("saving 1 job for sampling:", job_name + ".sh")
-                sample_job_flag = False
-            else:
-                print("for file explosion reasons, removing:", job_name +".sh") 
-                os.remove(os.path.join(gene_annotation_BLAT_path, job_name + ".sh"))
+                process.start()
+                mp_store.append(process)
                 
-            process.start()
-            mp_store.append(process)
+            
         for item in mp_store:
             item.join()
         mp_store[:] = []
-        write_to_bypass_log(output_folder_path, gene_annotation_BLAT_label)
-        
-   
-        
+        write_to_bypass_log(output_folder_path, gene_annotation_BLAT_cat_label)
     
     if check_bypass_log(output_folder_path, gene_annotation_BLAT_pp_label):
         
@@ -806,7 +825,7 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
                     Blat_pp_Pool.apply_async(commands.create_and_launch,
                         args=(
                             gene_annotation_BLAT_label,
-                            commands.create_BLAT_pp_command_v2(gene_annotation_BLAT_label, full_sample_path, fasta_db),
+                            commands.create_BLAT_pp_command_v2(gene_annotation_BLAT_label, full_sample_path, gene_annotation_BWA_label),
                             True,
                             job_name
                         )
@@ -834,25 +853,26 @@ def main(config_path, pair_1_path, pair_2_path, single_path, output_folder_path,
     GA_DIAMOND_tool_output_path = os.path.join(gene_annotation_DIAMOND_path, "data", "0_diamond")
     #if not check_where_resume(None, GA_DIAMOND_tool_output_path, gene_annotation_BLAT_path, file_check_bypass = True):
     if check_bypass_log(output_folder_path, gene_annotation_DIAMOND_label):
-        sections = ["contigs", "singletons"]
-        if read_mode == "paired":
-            sections.extend(["pair_1", "pair_2"])
-        for section in sections:
-            inner_name = section + "_run_diamond"
-            process = mp.Process(
-                target=commands.create_and_launch,
-                args=(
-                    gene_annotation_DIAMOND_label,
-                    commands.create_DIAMOND_annotate_command(gene_annotation_DIAMOND_label, gene_annotation_BLAT_label, section),
-                    True,
-                    inner_name
+        real_thread_count = thread_count
+        if(thread_count == 1):
+            real_thread_count = 2
+        DIAMOND_Pool = mp.Pool(int(real_thread_count / 2))
+        for item in os.listdir(os.path.join(gene_annotation_BLAT_path, "final_results")):
+            if(item.endswith(".fasta")):
+                file_tag = os.path.basename(split_sample)
+                file_tag = os.path.splitext(file_tag)[0]
+                job_name = "DIAMOND_" + file_tag
+                full_sample_path = os.path.join(os.path.join(gene_annotation_BLAT_path, "final_results", split_sample))
+                DIAMOND_Pool.apply_async(commands.create_and_launch,
+                    args=(
+                        gene_annotation_DIAMOND_label,
+                        commands.create_DIAMOND_annotate_command_v2(gene_annotation_DIAMOND_label, full_sample_path),
+                        True,
+                        job_name
+                    )
                 )
-            )
-            process.start()
-            mp_store.append(process)
-        for item in mp_store:
-            item.join()
-        mp_store[:] = []
+        DIAMOND_Pool.close()
+        DIAMOND_Pool.join()
         write_to_bypass_log(output_folder_path, gene_annotation_DIAMOND_label)
         
     #if not check_where_resume(gene_annotation_DIAMOND_path, None, GA_DIAMOND_tool_output_path, file_check_bypass = True):
